sage/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sage"
version = "0.1.0"
description = "Sage - Gemini 2.5 Flash Integration for System Intelligence"
authors = [
    {name = "coff33ninja", email = "coff33ninja69@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
keywords = ["ai", "gemini", "system-monitoring", "intelligence", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
]

dependencies = [
    "google-generativeai>=0.8.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
    "redis>=5.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
sage = "cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "*.egg-info"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
